Add shared etl functions
Functions used in multiple etl routines
are pulled out to shared module.
